🏗️  Complete user cases

+++ frontend/static/views/SelectView.js
@@ -1,7 +1,6 @@
 import AbstractView from "./AbstractView.js";
 import NavComponent from "../js/common/nav.js";
 import initSeats from "../js/seat-selection/seat-selection.js";
-import sessionStorageToObj from "../js/common/sessionStorageToObj.js";
 import toast from "../js/common/toast.js";
 
 const Price = {
@@ -48,16 +47,6 @@ export default class extends AbstractView {
 
           <main class="seat-view">
             <div id="disable--seat-view"></div>
-            <!-- <nav class="seat-view__nav">
-              <a href="/main" data-link>
-              <button type="button" class="btn btn-primary">
-                <div class="plus-icon">
-                  <img src="/static/assets/svg/arrow-left.svg" />
-                </div>
-                <a href='/paycheck' data-link id="prevBtn"><div class="select-label">이전페이지로</div></a>
-              </button>
-              </a>
-            </nav> -->
 
             <div class="information">
               <!-- <img src="/static/assets/images/table-ok.png" /> -->
@@ -94,6 +83,7 @@ export default class extends AbstractView {
     document.getElementById("root").appendChild(script);
 
     initSeats();
+    // countSelected = 0;
 
     const path = sessionStorage.getItem("path");
     const prevPath = sessionStorage.getItem("history");
@@ -107,8 +97,7 @@ export default class extends AbstractView {
     if (prevPath === "login") {
       prevBtn.setAttribute("href", "/ticket");
 
-      const ticketObj = sessionStorageToObj("ticket");
-      const { time, auth } = ticketObj;
+      const { time, auth } = JSON.parse(localStorage.getItem("ticket"));
       const formattedPrice = `${Intl.NumberFormat("ko-KR").format(
         Price[auth][time]
       )}원`;
@@ -122,8 +111,8 @@ export default class extends AbstractView {
         //선택한 좌석이 있을 경우
         if (!selectedSeat) {
           payBtn.parentElement.setAttribute("href", "/select");
-          // toast("좌석을 선택해주세요!");
-          alert("좌석을 선택해주세요");
+          toast("좌석을 선택해주세요!");
+          // alert("좌석을 선택해주세요");
         } else {
           //validation - 좌석 data fetch받아와서 자리 여전히 없으면 req 넘기고, 다음 페이지로 이동
           if (true) {
@@ -153,11 +142,49 @@ export default class extends AbstractView {
     }
     //퇴실 Main에서 버튼 선택
     else if (prevPath === "before") {
-      prevBtn.setAttribute("href", "/main"); //퇴실메인으로 표시해야 함
-      //좌석 선택 - 퇴실뷰 나오면 그때 구현!!!
+      prevBtn.setAttribute("href", "/main"); //이거 메인으로 이동안함?!?!
+
       if (path === "select") {
-        sessionStorage.clear();
-        sessionStorage.setItem("history", "main");
+        const $usingTime = document.getElementById("charged-time");
+        const paymethod = document.getElementById("pay-method");
+
+        payBtn.value = "좌석 이동하기";
+        totalPrice.innerText = `0원`;
+        $usingTime.innerText = "";
+        paymethod.innerText = "시간권 차감";
+
+        payBtn.addEventListener("click", (e) => {
+          const selectedSeat = sessionStorage.getItem("lastSelected");
+          //선택한 좌석이 있을 경우
+          if (!selectedSeat) {
+            payBtn.parentElement.setAttribute("href", "/select");
+            // toast("좌석을 선택해주세요!");
+            alert("좌석을 선택해주세요");
+          } else {
+            //validation - 좌석 data fetch받아와서 자리 여전히 없으면 req 넘기고, 다음 페이지로 이동
+            if (true) {
+              payBtn.parentElement.setAttribute("href", "/paycheck");
+              const req = {
+                category: auth,
+                duration: time,
+                price: Price[auth][time],
+                table: sessionStorage.getItem("table"),
+                position: Number(selectedSeat.replace(/[^0-9]/g, "")),
+                //현재 시각
+              };
+              //api.post(req)
+              console.log("selected seta", req);
+
+              sessionStorage.clear();
+              sessionStorage.setItem("history", "using");
+            } else {
+              alert("좌석을 다시 선택해주세요");
+              payBtn.parentElement.setAttribute("href", "/select");
+              localStorage.removeItem("table");
+              localStorage.removeItem("lastSelected");
+            }
+          }
+        });
       }
       //시간만 연장
       else if (path === "extend") {
@@ -165,8 +192,19 @@ export default class extends AbstractView {
         const warningMessage = document.createElement("h1");
         const ticketType = document.querySelector(".seat__selected");
 
-        seatTitle.innerText = "Your Selected Ticket";
-        ticketType.innerText = disableSelect.style.visibility = "visible";
+        const { time, auth } = JSON.parse(localStorage.getItem("ticket"));
+        localStorage.removeItem("ticket");
+        const formattedPrice = `${Intl.NumberFormat("ko-KR").format(
+          Price[auth][time]
+        )}원`;
+
+        ticketInfoArr[0].innerText = auth === "oneday" ? "당일권" : "시간권";
+        ticketInfoArr[1].innerText = `${time}시간`;
+        totalPrice.innerText = formattedPrice;
+
+        seatTitle.innerHTML = "Your <br>Selected<br> Ticket";
+        ticketType.innerText = ticketInfoArr[0].innerText;
+        disableSelect.style.visibility = "visible";
 
         warningMessage.innerText =
           "좌석 선택은 이용권 연장 완료 후 시도해 주시기 바랍니다.";
@@ -175,18 +213,19 @@ export default class extends AbstractView {
         warningMessage.style.padding = "2em";
         disableSelect.append(warningMessage);
 
-        prevBtn.addEventListener("click", () => {});
+        prevBtn.addEventListener("click", () => {
+          sessionStorage.setItem("path", "extend");
+        });
 
         sessionStorage.clear();
         sessionStorage.setItem("history", "before");
+        sessionStorage.setItem("price", formattedPrice);
       }
     }
     // 사용중 Main에서 버튼 선택
     else if (prevPath === "using") {
       // &&시간이 남아있는 경우에만 자리이동 가능(이용중 메인에서 온 경우)
       if (path === "move") {
-        // 왼편 결제하기 컴포넌트 지우기, 결제하기버튼 -> 이동하기 버튼으로 바꾸기
-        // const $paymentBtn = document.getElementById("payment-Btn");
         const $usingTime = document.getElementById("charged-time");
         const paymethod = document.getElementById("pay-method");
 
@@ -234,6 +273,15 @@ export default class extends AbstractView {
       else if (path == "extend") {
         const disableSelect = document.getElementById("disable--seat-view");
         const warningMessage = document.createElement("h1");
+
+        const { time, auth } = JSON.parse(localStorage.getItem("ticket"));
+        const formattedPrice = `${Intl.NumberFormat("ko-KR").format(
+          Price[auth][time]
+        )}원`;
+
+        ticketInfoArr[0].innerText = auth === "oneday" ? "당일권" : "시간권";
+        ticketInfoArr[1].innerText = `${time}시간`;
+        totalPrice.innerText = formattedPrice;
 
         seatTitle.innerText = "Your Selected Ticket";
         disableSelect.style.visibility = "visible";
